pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-checkpoint-amazon-dynamodb"
version = "0.1.2"
authors = [
    { name = "Aaron Su", email = "supaaron@amazon.com" },
]
description = "Library with a single-table DynamoDB implementation of LangGraph checkpoint saver."
readme = "README.md"
requires-python = ">=3.9.0,<4.0.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langgraph>=0.2.0",
    "aioboto3>=13.0.0",
]

[project.optional-dependencies]
infra = [
    "aws-cdk-lib>=2.174.0",
    "constructs>=10.0.0,<11.0.0",
]

[project.urls]
"Homepage" = "https://github.com/aaronsu11/langgraph-checkpoint-dynamodb"
"Bug Tracker" = "https://github.com/aaronsu11/langgraph-checkpoint-dynamodb/issues"